pyproject.toml
[project]
name = "voicecontrol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "just>=0.8.162",
    "langchain>=0.3.22",
    "langchain-ollama>=0.3.0",
    "loguru>=0.7.3",
    "opencv-python>=4.11.0.86",
    "playwright>=1.51.0",
    "psutil>=7.0.0",
    "pyautogui>=0.9.54",
    "pyyaml>=6.0.2",
    "ruff>=0.9.6",
    "toml>=0.10.2",
    "uvicorn>=0.34.0",
    "zmq>=0.0.0",
]

# ruff
[tool.ruff]
line-length = 120
exclude = [
    "build",
    "dist",
    "venv",
    ".tox",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
    ".vscode",
    ".idea",
    ".mypy_cache",
    ".pytest_cache",
    ".vscode",
    ".idea",
]

[tool.ruff.lint]
select = ["ALL"]
#"ANN001", "ANN002", "ANN003"